feat(plotting): matplotlib backend and backend= for summary_plot (#639)
Add MatplotlibBackend (seaborn styling/faceting helpers), resolve backends
via get_backend, deprecate summary_plot(interactive=) in favour of
backend=, and delete SeabornPlotBuilder.
