+++ .issueflows/01-current-issues/issue639_status.md
@@ -1,0 +1,14 @@
+# Issue #639 — Status
+
+- [ ] Done
+
+## What's done
+
+- Plan accepted (2026-07-23).
+- Branch `cursor/639-mpl-backend-4efd`; draft PR #645.
+
+## Remaining work
+
+- Add `MatplotlibBackend` (`backends/mpl.py`); `get_backend`.
+- Flip `summary_plot` to `backend=` + deprecate `interactive=`.
+- Delete `SeabornPlotBuilder`; update tests/oracle/DEPRECATIONS/design notes.
