+++ .issueflows/04-designs-and-guides/plotting-prepare.md
@@ -3,8 +3,8 @@
 ## Context
 
 `SummaryPlotDataPreparer` lived in `cellpy/utils/plotutils.py` and fed both
-`PlotlyPlotBuilder` and `SeabornPlotBuilder`. Epic #567 Stage 1 needs
-prepare → `FigureSpec` → backend.render as the only summary path.
+plotly and seaborn builders. Epic #567 Stage 1 needs prepare → `FigureSpec` →
+backend.render as the only summary path.
 
 ## Decision
 
@@ -12,15 +12,16 @@ prepare → `FigureSpec` → backend.render as the only summary path.
   `prepare(ctx, family, config) -> (frame, FigureSpec)`.
 - **`SummaryPlotDataPreparer`** moved there (implementation detail); deleted
   from `plotutils`.
-- **`FigureSpec.extras`** carries `prepared_data_info` (seaborn bridge) and
-  `render` (plotly knobs including precomputed formation / no-formation
-  layout). First-class panel/axis fields grow as later issues need them.
+- **`FigureSpec.extras`** carries `prepared_data_info` and `render` (plotly
+  knobs including precomputed formation / no-formation layout). For the
+  matplotlib backend, `summary_plot` also attaches live `config` / `cell` on
+  `extras` so seaborn styling knobs remain available without a second prepare.
 - **`CellContext`** in `cellpy/plotting/context.py` is the thin cell adapter;
   BatchContext waits for collectors rebase.
-- Public `summary_plot` stays in `plotutils` but only orchestrates
-  context → registry → prepare → render.
+- Public `summary_plot` stays in `plotutils` and orchestrates
+  context → registry → prepare → `get_backend(backend).render`.
 
 ## Links
 
-- Issue #638; epic #567
+- Issues #639, #638; epic #567
 - Related: `plotting-registry.md`, `plotting-backends.md`
